Split code emitter over multiple files
